requirements.txt
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
flask==3.0.0
tmdbv3api==1.9.0
requests==2.31.0
python-dotenv==1.0.0
Flask-Caching==2.1.0
